+++ sync-for-expenses/RELEASES.md
@@ -709,3 +709,13 @@ Based on:
 - [typescript v4.0.0] sync-for-expenses
 ### Releases
 - [NPM v4.0.0] https://www.npmjs.com/package/@codat/sync-for-expenses/v/4.0.0 - sync-for-expenses
+
+## 2023-12-11 01:30:12
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses.yaml
+- Speakeasy CLI 1.125.2 (2.210.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.1] sync-for-expenses
+### Releases
+- [NPM v4.0.1] https://www.npmjs.com/package/@codat/sync-for-expenses/v/4.0.1 - sync-for-expenses